requirements.txt
aiosqlite==0.19.0
alembic==1.10.3
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post1
fastapi==0.95.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Mako==1.2.4
MarkupSafe==2.1.2
orjson==3.8.10
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
